requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
python-dotenv
